PC-1085: split AdminAction
introduce a new CommandHandler class that does the printing and has a public method for each command

AdminAction now interfaces with the objects and calls database methods
